+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 2.7.3 (TBD)
+
+- Bugfix: Standard I/O is restored when using `telepresence intercept <opts> -- <command>`.
+
+- Bugfix: Graciously handle nil intercept environment from the traffic-manager.
+
 ### 2.7.2 (TBD)
 
 - Feature: The timeout for the initial connectivity check that Telepresence performs